+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 - Support key import from PEM files ([#99](https://github.com/Nitrokey/nethsm-sdk-py/issues/99))
+- Add `ignore_whitespace` option to `Base64.from_encoded` ([#108](https://github.com/Nitrokey/nethsm-sdk-py/issues/108))
 
 [All Changes](https://github.com/Nitrokey/nethsm-sdk-py/compare/v0.5.0...HEAD)